fds-research/requirements.txt
pandas>=2.0
scikit-learn>=1.3
lightgbm>=4.0
PyYAML>=6.0
joblib>=1.3
